Refactor code and finish Discovery's tests
With the additional parameters for the Discovery, card and also for
the whole of the Maestro Card, a more dynamic approach was required.
The tests for the Discover card are based off of data in 2 arrays,
as opposed to being hard coded for each test. After doing that, it
was clear that this approach should be brought to the
detectNetwork.js file as well.

Now detectNetwork.js has a cards array, filled with card objects, and
it goes through loops instead of a single switch statement, making
it more readable (at the cost of speed), and reducing the chance that
an error will go unnoticed.